# docs/openapi/agent-platform-openapi.json
{
  "components": {
    "schemas": {
...
      "ApiChatResponse": {
        "properties": {
          "memory_diff": {
            "additionalProperties": true,
            "title": "Memory Diff",
            "type": "object"
          },
          "sequence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sequence",
            "type": "array"
          },
          "total_steps": {
            "default": 0,
            "title": "Total Steps",
            "type": "integer"
          }
        },
        "title": "ApiChatResponse",
        "type": "object"
      },
      "ApiCommentingGenerateResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "content_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Source"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason"
          },
          "max_tokens": {
            "title": "Max Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "model_key": {
            "title": "Model Key",
            "type": "string"
          },
          "persona_key": {
            "title": "Persona Key",
            "type": "string"
          },
          "prompt_key": {
            "title": "Prompt Key",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "provider_model_id": {
            "title": "Provider Model Id",
            "type": "string"
          },
          "raw_reply": {
            "additionalProperties": true,
            "title": "Raw Reply",
            "type": "object"
          },
          "raw_request": {
            "additionalProperties": true,
            "title": "Raw Request",
            "type": "object"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "scenario": {
            "enum": [
              "chat",
              "comment",
              "label"
            ],
            "title": "Scenario",
            "type": "string"
          },
          "selected_attempt": {
            "title": "Selected Attempt",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          },
          "task_shape": {
            "enum": [
              "classic",
              "all_in_system",
              "structured_output"
            ],
            "title": "Task Shape",
            "type": "string"
          },
          "timeout_seconds": {
            "title": "Timeout Seconds",
            "type": "number"
          },
          "usage": {
            "additionalProperties": true,
            "title": "Usage",
            "type": "object"
          }
        },
        "required": [
          "scenario",
          "model_key",
          "source_id",
          "source_label",
          "provider_model_id",
          "prompt_key",
          "persona_key",
          "model",
          "content",
          "provider",
          "max_tokens",
          "timeout_seconds",
          "task_shape",
          "selected_attempt"
        ],
        "title": "ApiCommentingGenerateResponse",
        "type": "object"
      },
...
      "ChatRequest": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retry_count": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Count"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 600.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "agent_id",
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "CommentingGenerateRequest": {
        "examples": [
          {
            "input": "Summarize the reader reaction and write one concise editor-style reply.",
            "max_tokens": 512,
            "model_key": "local_llama_server::gemma4",
            "persona_key": "comment_linxiaotang",
            "prompt_key": "comment_v20260418",
            "retry_count": 1,
            "scenario": "comment",
            "task_shape": "classic",
            "timeout_seconds": 120
          }
        ],
        "properties": {
          "input": {
            "description": "News article, comment thread, or source text to comment on.",
            "title": "Input",
            "type": "string"
          },
          "max_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional response token budget. Defaults to Comment Lab runtime settings.",
            "title": "Max Tokens"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "deprecated": true,
            "description": "Legacy selector kept for backward compatibility. Prefer `model_key`.",
            "title": "Model"
          },
          "model_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Router-scoped model key from `/api/v1/options?scenario=comment`, for example `local_llama_server::gemma4`.",
            "examples": [
              "local_llama_server::gemma4"
            ],
            "title": "Model Key"
          },
          "persona_key": {
            "default": "comment_linxiaotang",
            "description": "Comment Lab persona key from `/api/v1/options?scenario=comment`.",
            "title": "Persona Key",
            "type": "string"
          },
          "prompt_key": {
            "default": "comment_v20260418",
            "description": "Comment Lab prompt key from `/api/v1/options?scenario=comment`.",
            "title": "Prompt Key",
            "type": "string"
          },
          "retry_count": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional provider retry count for transient failures.",
            "title": "Retry Count"
          },
          "scenario": {
            "default": "comment",
            "description": "Must be `comment` for this endpoint.",
            "enum": [
              "chat",
              "comment",
              "label"
            ],
            "title": "Scenario",
            "type": "string"
          },
          "task_shape": {
            "anyOf": [
              {
                "enum": [
                  "classic",
                  "all_in_system",
                  "structured_output"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Prompt-packing strategy. Defaults to the Comment Lab runtime setting.",
            "title": "Task Shape"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional provider timeout in seconds. Use a realistic local-model value such as 120.",
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "input"
        ],
        "title": "CommentingGenerateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LabelingGenerateRequest": {
        "examples": [
          {
            "input": "Messi scored for Inter Miami against Orlando City.",
            "max_tokens": 1024,
            "model_key": "local_llama_server::gemma4",
            "prompt_key": "label_football_entities_v1",
            "repair_retry_count": 1,
            "scenario": "label",
            "schema_key": "label_football_entity_groups_v1",
            "timeout_seconds": 120
          }
        ],
        "properties": {
          "input": {
            "description": "Article or text to extract grouped entity lists from.",
            "title": "Input",
            "type": "string"
          },
          "max_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional response token budget. Defaults to Label Lab runtime settings.",
            "title": "Max Tokens"
          },
          "model_key": {
            "description": "Router-scoped model key from `/api/v1/options?scenario=label`, for example `local_llama_server::gemma4`.",
            "examples": [
              "local_llama_server::gemma4"
            ],
            "title": "Model Key",
            "type": "string"
          },
          "prompt_key": {
            "default": "label_generic_entities_v1",
            "description": "Label Lab prompt key from `/api/v1/options?scenario=label`.",
            "title": "Prompt Key",
            "type": "string"
          },
          "repair_retry_count": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Number of structured-output repair attempts after validation failure.",
            "title": "Repair Retry Count"
          },
          "scenario": {
            "default": "label",
            "description": "Must be `label` for this endpoint.",
            "enum": [
              "chat",
              "comment",
              "label"
            ],
            "title": "Scenario",
            "type": "string"
          },
          "schema_key": {
            "default": "label_entity_groups_v1",
            "description": "Label Schema Center key from `/api/v1/options?scenario=label`.",
            "title": "Schema Key",
            "type": "string"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional provider timeout in seconds. Use a realistic local-model value such as 120.",
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "input",
          "model_key"
        ],
        "title": "LabelingGenerateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Provides versioned API routes for Agent Platform runtime/control/test orchestration. Designed for backend-first API consumption and ADE frontend integration.",
    "summary": "Runtime and control APIs for ADE and local Agent Platform workflows",
    "title": "Agent Platform API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "operationId": "api_chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send a chat message to a persistent Agent Studio agent",
        "tags": [
          "Agent Studio"
        ]
      }
    },
    "/api/v1/commenting/generate": {
      "post": {
        "operationId": "api_commenting_generate_api_v1_commenting_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "llama_server_comment": {
                  "summary": "Generate with local llama-server",
                  "value": {
                    "input": "Summarize the reader reaction and write one concise editor-style reply.",
                    "max_tokens": 512,
                    "model_key": "local_llama_server::gemma4",
                    "persona_key": "comment_linxiaotang",
                    "prompt_key": "comment_v20260418",
                    "retry_count": 1,
                    "scenario": "comment",
                    "task_shape": "classic",
                    "timeout_seconds": 120
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CommentingGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiCommentingGenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate a stateless comment for news/comment threads",
        "tags": [
          "Comment Lab"
        ]
      }
    },
    "/api/v1/labeling/generate": {
      "post": {
        "operationId": "api_labeling_generate_api_v1_labeling_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "football_entities": {
                  "summary": "Extract football players and teams",
                  "value": {
                    "input": "Messi scored for Inter Miami against Orlando City.",
                    "max_tokens": 1024,
                    "model_key": "local_llama_server::gemma4",
                    "prompt_key": "label_football_entities_v1",
                    "repair_retry_count": 1,
                    "scenario": "label",
                    "schema_key": "label_football_entity_groups_v1",
                    "timeout_seconds": 120
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/LabelingGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLabelingGenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate stateless grouped entity extraction for an input article",
        "tags": [
          "Label Lab"
        ]
      }
    },
    "/api/v1/options": {
      "get": {
        "operationId": "api_get_options_api_v1_options_get",
        "parameters": [
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "scenario",
            "required": false,
            "schema": {
              "default": "chat",
              "title": "Scenario",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOptionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List runtime options for an ADE scenario",
        "tags": [
          "Platform Meta"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Agent Platform API local",
      "url": "http://127.0.0.1:8284"
    }
  ],
...
}